src/main/java/nl/ivonet/comics/directory/ExtensionFilter.java
/*
 * Copyright (c) 2013 by Ivo Woltring (http://ivonet.nl)
 *
 *   Licensed under the Apache License, Version 2.0 (the "License");
 *   you may not use this file except in compliance with the License.
 *   You may obtain a copy of the License at
 *
 *          http://www.apache.org/licenses/LICENSE-2.0
 *
 *   Unless required by applicable law or agreed to in writing, software
 *   distributed under the License is distributed on an "AS IS" BASIS,
 *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *   See the License for the specific language governing permissions and
 *   limitations under the License.
 */

package nl.ivonet.comics.directory;

import nl.ivonet.cdi_properties.Property;

import javax.inject.Inject;
import java.io.IOException;
import java.nio.file.DirectoryStream;
import java.nio.file.Files;
import java.nio.file.Path;
import java.util.Locale;
import java.util.stream.Stream;

/**
 * @author Ivo Woltring
 */
public class ExtensionFilter implements DirectoryStream.Filter<Path> {

    private static final String DELIMETER = ":";

    @Inject @Property private String filterExtensions;

    @Override
    public boolean accept(final Path entry) throws IOException {
        return Files.isRegularFile(entry) && isAcceptable(entry);
    }

    private boolean isAcceptable(final Path entry) {
        final String[] split = this.filterExtensions.toLowerCase(Locale.US)
                                                    .split(DELIMETER);
        return Stream.of(split)
                     .anyMatch(p -> entry.getFileName()
                                         .toString()
                                         .toLowerCase(Locale.US)
                                         .endsWith(p));
    }
}
